\docType{data}
\name{andalucia-data}
\alias{andalucia-data}
\alias{andalucia}
\title{Ciudades de Andalucia}
\format{\code{name}: El nombre de una ciudad o de una etiqueta relativa de posicion.

\code{x}: La coordenada x de los puntos.

\code{y}: La coordenada y de los puntos

\code{city}: Si yes entonces el punto es una ciudad, en otro caso es un limite.}
\description{
El conjunto de datos 'andalusia' tiene 12 filas y 4 columnas, que son la
posicion geografica de las capitales de provincia andaluzas.
}
\section{Uso}{

data('andalusia')
}

\section{Fuente}{
Los datos se han tomado de wikipedia.
}

\seealso{
Vease tambien \code{\link{orloca-package}}.
}
\keyword{Andalucia}
\keyword{data}
