\docType{data}
\name{andalusia}
\alias{andalucia}
\alias{andalusia}
\encoding{UTF-8}

\title{Ciudades de Andalucia}
\format{\code{name}: El nombre de una ciudad o de una etiqueta relativa de posici\enc{ó}{o}n

\code{x}: La coordenada x de los puntos

\code{y}: La coordenada y de los puntos

\code{city}: Si yes entonces el punto es una ciudad, en otro caso es un l\enc{í}{i}mite.}
\description{
El conjunto de datos 'andalusia' tiene 12 filas y 4 columnas, que son la
posici\enc{ó}{o}n geogr\enc{á}{a}fica de las capitales de provincia andaluzas.
}
\section{Uso}{

data('andalusia')
}

\section{Fuente}{
Los datos se han tomado de wikipedia.
}

\seealso{
V\enc{é}{e}ase tambi\enc{é}{e}n \code{\link{orloca-package}}.
}
\keyword{Andalucia}
\keyword{Andalusia}
\keyword{data}
