\name{andalusia-data}
\docType{data}
\alias{andalusia}
\alias{andalucia}
\encoding{UTF-8}

\title{Ciudades de Andalucia}
\description{El conjunto de datos 'andalusia' tiene 12 filas y 4
  columnas, con las coordenadas \enc{geográficas}{geograficas} de las capitales de
  provincia andaluzas.}
\usage{andalusia}
\section{Formato}{
  \code{name}: El nombre de las ciudades o una etiqueta de \enc{posición}{posicion} relativa.
  
  \code{x}: La coordenada x de los puntos.
  
  \code{y}: La coordenada y de los puntos.
  
  \code{city}: 'yes' significa que es una ciudad, en caso contrario es
  un \enc{límite}{limite}.
}
\source{Los datos se han tomado de wikipedia.}
\keyword{data}
\keyword{andalusia}
\keyword{andalucia}
\seealso{See also \code{\link{orloca.es-package}}.}