\docType{methods}
\encoding{UTF-8}
\name{zsumgra}
\alias{zsumgra}
\title{zsumgra}
\usage{
zsumgra(...)
}
\arguments{
\item{\ldots}{Par\enc{á}{a}metros pasados a distsumgra}
}
\description{
La funci\enc{ó}{o}n zsumgra est\enc{a}{a} obsoleta y podr\enc{í}{i}a ser suprimida en sucesivas versiones del paquete. Use \link{distsumgra} en su lugar.
}
\keyword{obsoleta}