\docType{methods}
\encoding{UTF-8}
\name{zsum}
\alias{zsum}
\title{zsum}
\usage{
zsum(...)
}
\arguments{
\item{\ldots}{Par\enc{á}{a}metros pasados a distsum}
}
\description{
La funci\enc{ó}{o}n zsum est\enc{a}{a} obsoleta y podr\enc{í}{i}a ser suprimida en sucesivas versiones del paquete. Use \link{distsum} en su lugar.
}
\keyword{obsoleta}