\docType{methods}
\encoding{UTF-8}
\name{zsummin}
\alias{zsummin}
\title{zsummin}
\usage{
zsummin(...)
}
\arguments{
\item{\ldots}{Par\enc{á}{a}metros pasados a distsummin}
}
\description{
La funci\enc{ó}{o}n zsummin est\enc{a}{a} obsoleta y podr\enc{í}{i}a ser suprimida en sucesivas versiones del paquete. Use \link{distsummin} en su lugar.
}
\keyword{obsoleta}